fix prompt logic
