+++ apps/sanity/CLI/fetchSanityOrganizations.mjs
