+++ apps/sanity/CLI/services/vercel.mjs
@@ -31,6 +31,58 @@ export async function getVercelTeams() {
   }));
 }
 
+export async function getVercelUserInfo() {
+  const envs = loadEnvVariables();
+  const token = envs.VERCEL_PERSONAL_AUTH_TOKEN;
+
+  const response = await fetch("https://api.vercel.com/v2/user", {
+    method: "GET",
+    headers: {
+      "Content-Type": "application/json",
+      Authorization: `Bearer ${token}`,
+    },
+  });
+
+  if (!response.ok) {
+    console.log(response.status, response.statusText, await response.json());
+    throw new Error(`❌ HTTP error! Status: ${response.status}`);
+  }
+
+  const data = await response.json();
+
+  return {
+    name: data.user.name,
+    email: data.user.email,
+    profileImage: data.user.profileImage,
+  };
+}
+
+export async function getVercelProjects() {
+  const envs = loadEnvVariables();
+  const repoId = envs.REPO_ID;
+  const vercelTeamId = envs.VERCEL_FR_TEAM_ID;
+  const vercelToken = envs.VERCEL_PERSONAL_AUTH_TOKEN;
+
+  const response = await fetch(
+    `https://api.vercel.com/v9/projects?repoId=${repoId}&teamId=${vercelTeamId}`,
+    {
+      method: "GET",
+      headers: {
+        "Content-Type": "application/json",
+        Authorization: `Bearer ${vercelToken}`,
+      },
+    },
+  );
+
+  if (!response.ok) {
+    throw new Error(`❌ HTTP error! Status: ${response.status}`);
+  }
+
+  const data = await response.json();
+
+  return data.projects;
+}
+
 export async function createVercelProject({ projectName, sbParams }) {
   const { isPreview, storyblokToken, whRevalidateSecret } = sbParams;
   const envs = loadEnvVariables();
@@ -103,32 +155,6 @@ export async function createVercelProject({ projectName, sbParams }) {
     projectName: data.name,
     deploymentUrl: `https://${data.name}.vercel.app`,
   };
-}
-
-export async function getVercelProjects() {
-  const envs = loadEnvVariables();
-  const repoId = envs.REPO_ID;
-  const vercelTeamId = envs.VERCEL_FR_TEAM_ID;
-  const vercelToken = envs.VERCEL_PERSONAL_AUTH_TOKEN;
-
-  const response = await fetch(
-    `https://api.vercel.com/v9/projects?repoId=${repoId}&teamId=${vercelTeamId}`,
-    {
-      method: "GET",
-      headers: {
-        "Content-Type": "application/json",
-        Authorization: `Bearer ${vercelToken}`,
-      },
-    },
-  );
-
-  if (!response.ok) {
-    throw new Error(`❌ HTTP error! Status: ${response.status}`);
-  }
-
-  const data = await response.json();
-
-  return data.projects;
 }
 
 export async function createProjectDeployment({ name, id }) {
