+++ apps/sanity/sanity.cli.ts
@@ -1,7 +1,7 @@
 // import { loadEnvConfig } from "@next/env";
 import { defineCliConfig } from "sanity/cli";
 
-console.log("add logic to load ebv variables");
+// console.log("add logic to load ebv variables");
 
 // const dev = process.env.NODE_ENV !== "production";
 // loadEnvConfig(__dirname, dev, { info: () => null, error: console.error });
